only build linkvertise on firefox
