+++ scripts/build_js/firefox.js
@@ -1,3 +1,5 @@
+const fs = require('fs').promises;
+
 if (require.main === module)
     throw new Error('This file is a build module, run the build.js file to actually build FastForward');
 
@@ -6,6 +8,12 @@ const zipper = require('adm-zip');
 module.exports = async function ({versioning, destination, commit_number, version} = {}) {
     console.log(`[FastForward.build.firefox] building the FireFox package`);
     const z = new zipper();
+
+    console.log(`[FastForward.build.firefox] injecting the linkvertise bypass`);
+
+    const bypass = await fs.readFile(`${process.cwd()}/src/linkvertise.js`);
+
+    await fs.appendFile(`${process.cwd()}/build/FastForward.firefox/injection_script.js`, bypass);
 
     z.addLocalFolder(`${process.cwd()}/${destination}`);
